update some train_fn to train_fn_ref
