+++ python/ray/train/v2/api/data_parallel_trainer.py
@@ -43,7 +43,7 @@ from ray.train.v2._internal.execution.context import TrainRunContext
 from ray.train.v2._internal.execution.controller import TrainController
 from ray.train.v2._internal.execution.failure_handling import create_failure_policy
 from ray.train.v2._internal.execution.scaling_policy import create_scaling_policy
-from ray.train.v2._internal.util import construct_train_func
+from ray.train.v2._internal.util import construct_train_func, ObjectRefWrapper
 from ray.train.v2.api.callback import UserCallback
 from ray.util.annotations import Deprecated, DeveloperAPI
 from ray.util.scheduling_strategies import NodeAffinitySchedulingStrategy
@@ -110,9 +110,10 @@ class DataParallelTrainer:
             train_func_context=self.backend_config.train_func_context,
             fn_arg_name="train_loop_per_worker",
         )
+        train_fn_ref = ObjectRefWrapper(train_fn)
 
         result = self._initialize_and_run_controller(
-            train_fn=train_fn,
+            train_fn_ref=train_fn_ref,
             scaling_policy=create_scaling_policy(self.scaling_config),
             failure_policy=create_failure_policy(self.run_config.failure_config),
             train_run_context=self.train_run_context,
